API Reference

PiCraft.turtle

PiCraft.clearEvents

PiCraft.connectToWorld

PiCraft.drawLine

PiCraft.getBlock

PiCraft.getHeight

PiCraft.getPlayerIds

PiCraft.getPos

PiCraft.getPos

PiCraft.getTile

PiCraft.getTile

PiCraft.mc_send

PiCraft.move

PiCraft.pitch

PiCraft.pollBlockHits

PiCraft.post

PiCraft.restoreWorld

PiCraft.roll

PiCraft.saveWorld

PiCraft.setBlock

PiCraft.setBlocks

PiCraft.setPos

PiCraft.setPos

PiCraft.setTile

PiCraft.setTile

PiCraft.yaw

PiCraft.turtle — Type.

turtle

A graphics turtle meant to move in all 3-Dimensions. It contains the following fields:

pos : The coordinates of the turtle.

direction : The Roll Axis - direction in which the turtle is facing.

normal : The Yaw axis - direction normal to the turtle's body.

penBlock : The block which will be used to draw.

stepSize : The size of a turtle's step.

penDown : Activation state of penBlock.

The default constructor initializes the turtle at the player's position,roll axis is the positive x-direction, with the yaw axis pointing downwards. penBlock is gold by default and is activated with the stepSize = 0.5

source

PiCraft.clearEvents — Method.

Clear all events in the buffer

source

PiCraft.connectToWorld — Function.

connectToWorld(address = "localhost", port = 4711)

Connect to the Minecraft API.

source

PiCraft.drawLine — Function.

drawLine(p1::Tuple{Real,Real,Real}, p2::Tuple{Real,Real,Real}, block::PiCraft.Block = PiCraft.GOLD_BLOCK)

Draw a line from point p1 to point p2 using block. block is gold block by default.

source

PiCraft.getBlock — Method.

getBlock(pos::Tuple{Real, Real, Real})

Get the Block information from the specified coordinates.

source

PiCraft.getHeight — Method.

getHeight(x::Int, z::Int)

Get the height of the world at the specified x and z coordinates.

source

PiCraft.getPlayerIds — Method.

getPlayerIds()

Return an array of all Player Id's connected to the server.

source

PiCraft.getPos — Method.

getPos(entityId::Int)

Get the position of the specified entity.

source

PiCraft.getPos — Method.

getPos()

Return the player's coordinates.

source

PiCraft.getTile — Method.

getTile(entityId::Int)

Get the tile on which the specified entity is.

source

PiCraft.getTile — Method.

getTile()

Return the tile's coordinates on which the player is standing.

source

PiCraft.mc_send — Function.

mc_send(cmd, output=true)

Communicate with the Minecraft API.

source

PiCraft.move — Method.

move(t::turtle, s::Int)

Move the turtle t forward s units.

source

PiCraft.pitch — Method.

pitch(t::turtle, θ::Real)

Pitch the turtle t by θ degrees.

source

PiCraft.pollBlockHits — Method.

pollBlockHits()

Returns an array of all the events which have occurred since the last time the function was called. Each event is described with a tuple ((x, y, z), face, entityId). x, y and z are the coordinates of the block. face is the block's face number which was hit and entityId identifies the player who hit the block using a sword.

source

PiCraft.post — Method.

post(s::String)

Post a message to chat

source

PiCraft.restoreWorld — Method.

Restore the world to the previous savepoint.

source

PiCraft.roll — Method.

roll(t::turtle, θ::Real)

Roll the turtle t by θ degrees.

source

PiCraft.saveWorld — Method.

Save the World's progress

source

PiCraft.setBlock — Method.

setBlock(pos::Tuple{Real, Real, Real}, block::Block)

Place the specified Block at the given coordinates.

source

PiCraft.setBlocks — Method.

setBlocks(p1::Tuple{Real, Real, Real}, p2::Tuple{Real, Real, Real}, block::Block)

Set an entire region to the specified block type defined by the corners.

source

PiCraft.setPos — Method.

setPos(entityId::Int, pos::Tuple{Real, Real, Real})

Set the position of the specified entity.

source

PiCraft.setPos — Method.

setPos(pos::Tuple{Real, Real, Real})

Teleport the player to the specified coordinates

source

PiCraft.setTile — Method.

setTile(entityId::Int)

Teleport the entity on the specified tile.

source

PiCraft.setTile — Method.

setTile(pos::Tuple{Real, Real, Real})

Teleport the player on top of the specified tile

source

PiCraft.yaw — Method.

yaw(t::turtle, θ::Real)

Yaw the turtle t by θ degrees.

source

getBlock(pos::Tuple{Real, Real, Real})

setBlock(pos::Tuple{Real, Real, Real}, block::Block)

setBlocks(p1::Tuple{Real, Real, Real}, p2::Tuple{Real, Real, Real}, block::Block)

getHeight(x::Int, z::Int)

getPlayerIds()

setting(name, status)

saveWorld()

restoreWorld()

post(s::string)

getTile()

setTile(pos::Tuple{Real, Real, Real})

getPos()

setPos(pos::Tuple{Real, Real, Real})